rename and rework identified bots examples
