+++ src/content/docs/bot-protection/reference/nextjs.mdx
@@ -25,10 +25,10 @@ import DecisionLogAppTS from "/src/content/docs/bot-protection/reference/snippet
 import DecisionLogPagesTS from "/src/content/docs/bot-protection/reference/snippets/nextjs/DecisionLogPages.ts?raw";
 import DecisionLogAppJS from "/src/content/docs/bot-protection/reference/snippets/nextjs/DecisionLogApp.js?raw";
 import DecisionLogPagesJS from "/src/content/docs/bot-protection/reference/snippets/nextjs/DecisionLogPages.js?raw";
-import BotTypeAppTS from "/src/content/docs/bot-protection/reference/snippets/nextjs/BotTypeApp.ts?raw";
-import BotTypePagesTS from "/src/content/docs/bot-protection/reference/snippets/nextjs/BotTypePages.ts?raw";
-import BotTypeAppJS from "/src/content/docs/bot-protection/reference/snippets/nextjs/BotTypeApp.js?raw";
-import BotTypePagesJS from "/src/content/docs/bot-protection/reference/snippets/nextjs/BotTypePages.js?raw";
+import IdentifiedBotsAppTS from "/src/content/docs/bot-protection/reference/snippets/nextjs/IdentifiedBotsApp.ts?raw";
+import IdentifiedBotsPagesTS from "/src/content/docs/bot-protection/reference/snippets/nextjs/IdentifiedBotsPages.ts?raw";
+import IdentifiedBotsAppJS from "/src/content/docs/bot-protection/reference/snippets/nextjs/IdentifiedBotsApp.js?raw";
+import IdentifiedBotsPagesJS from "/src/content/docs/bot-protection/reference/snippets/nextjs/IdentifiedBotsPages.js?raw";
 import ErrorsAppTS from "/src/content/docs/bot-protection/reference/snippets/nextjs/ErrorsApp.ts?raw";
 import ErrorsPagesTS from "/src/content/docs/bot-protection/reference/snippets/nextjs/ErrorsPages.ts?raw";
 import ErrorsAppJS from "/src/content/docs/bot-protection/reference/snippets/nextjs/ErrorsApp.js?raw";
@@ -348,25 +348,25 @@ be identified as zero, one, or more bots—all of which will be available on the
 <TabItem label="TS (App)">
 Create a new API route at `/app/api/route/hello.ts`:
 
-<Code code={BotTypeAppTS} title="/app/api/route/hello.ts" lang="ts" />
+<Code code={IdentifiedBotsAppTS} title="/app/api/route/hello.ts" lang="ts" />
 
 </TabItem>
 <TabItem label="TS (Pages)">
 Create a new API route at `/pages/api/hello.ts`:
 
-<Code code={BotTypePagesTS} title="/pages/api/hello.ts" lang="ts" />
+<Code code={IdentifiedBotsPagesTS} title="/pages/api/hello.ts" lang="ts" />
 
 </TabItem>
 <TabItem label="JS (App Router)">
 Create a new API route at `/app/api/arcjet/route.js`:
 
-<Code code={BotTypeAppJS} title="/app/api/route/hello.js" lang="js" />
+<Code code={IdentifiedBotsAppJS} title="/app/api/route/hello.js" lang="js" />
 
 </TabItem>
 <TabItem label="JS (Pages Router)">
 Create a new API route at `/pages/api/arcjet.js`:
 
-<Code code={BotTypePagesJS} title="/pages/api/hello.js" lang="js" />
+<Code code={IdentifiedBotsPagesJS} title="/pages/api/hello.js" lang="js" />
 
 </TabItem>
 </Tabs>
