+++ src/content/docs/bot-protection/reference/snippets/nextjs/IdentifiedBotsApp.js
@@ -1,5 +1,5 @@
 import arcjet, { detectBot } from "@arcjet/next";
-import type { NextApiRequest, NextApiResponse } from "next";
+import { NextResponse } from "next/server";
 
 const aj = arcjet({
   key: process.env.ARCJET_KEY,
@@ -11,21 +11,19 @@ const aj = arcjet({
   ],
 });
 
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse,
-) {
+export async function POST(req) {
   const decision = await aj.protect(req);
-  console.log("Decision", decision);
 
-  if (decision.isDenied() && decision.reason.isBot()) {
+  if (decision.reason.isBot()) {
     console.log("detected + allowed bots", decision.reason.allowed);
     console.log("detected + denied bots", decision.reason.denied);
-
-    return res.status(403).json({
-      error: "Forbidden",
-    });
   }
 
-  res.status(200).json({ name: "Hello world" });
+  if (decision.isDenied()) {
+    return NextResponse.json({ error: "You are a bot!" }, { status: 403 });
+  }
+
+  return NextResponse.json({
+    message: "Hello world",
+  });
 }
